refactor: 가독성을 위한 변수명 및 함수 리팩터
