+++ src/main/java/lotto/Application.java
@@ -1,7 +1,6 @@
 package lotto;
 
 import lotto.controller.LottoController;
-import lotto.parser.OutputParser;
 import lotto.view.InputView;
 import lotto.view.OutputView;
