+++ src/main/java/lotto/domain/Bonus.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import java.util.List;
-
 public class Bonus {
     private final int bonus;
