+++ src/main/java/lotto/factory/LottoFactory.java
@@ -1,7 +1,6 @@
 package lotto.factory;
 
 import camp.nextstep.edu.missionutils.Randoms;
-import java.util.ArrayList;
 import java.util.List;
 import lotto.domain.Lotto;
