+++ src/main/java/lotto/factory/LottosFactory.java
@@ -1,6 +1,5 @@
 package lotto.factory;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.IntStream;
 import lotto.domain.Lotto;
