+++ src/main/java/lotto/view/InputView.java
@@ -2,7 +2,6 @@ package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.List;
-import lotto.parser.InputParser;
 
 public class InputView {
     public int inputLottoPurchaseAmount() {
