+++ src/test/java/lotto/LottosRankResultTest.java
@@ -10,6 +10,7 @@ import lotto.domain.WinningLotto;
 import lotto.factory.LottoFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
 import static org.assertj.core.api.Assertions.*;
 
 public class LottosRankResultTest {
